requirements.txt
fastapi
uvicorn
supabase
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
aiosmtplib
email-validator
PyJWT
python-dateutil
bcrypt==4.1.2
python-dotenv
requests
langchain
langchain-groq
langgraph
langchain-community
langgraph-checkpoint-redis
redis
aiohttp
loguru
tqdm
duckduckgo-search